Created bit packer.
Accepts ASCII numbers (one per line), and will pack them into the smallest
amount of space.  We assume the biggest number we can receive first is 255.
Each number will increment an internal counter so we always know the
fewest bits we can use.

This is tested as working with a simple 1 - 10 value input.

I should add a test to make sure the input number is smaller than the biggest
expected number.